feat(transfuser): 切换至dino_geometry_openloop模型版本并调整相关参数
- 更新launch.json中model_version为'dino_geometry_openloop'
- 将num_frames参数从4调整为3，配合新模型设置
- 修改checkpoint路径和trajectory模式为'refine'
- 实现position_getter支持自定义时间列表以改进时空编码
- 调整TemporalWorldModel中的query_pos和keyval_pos调用，传入自定义时间列表
- 在transfuser_agent中新增加载'dino_geometry_openloop'模型逻辑
- 修改roPE频率参数，调整temporal和view频率值
- 修正agent_lightning_module中numpy转换数据类型为float以兼容新的输出格式
- 更新训练和评估脚本中相关路径、配置与参数，适配dino_geometry_openloop模型
- 降低训练脚本中的batch_size以适应新模型需求
- 延后curriculum学习阶段起止epoch号，调整训练计划
- 统一trainer中precision参数为'bf16-mixed'提高计算效率

+++ navsim/planning/training/agent_lightning_module.py
@@ -258,9 +258,9 @@ class AgentLightningModule(pl.LightningModule):
                 else:
                     prediction = self.agent.forward_test(features)
 
-                poses = prediction['trajectory'].cpu().numpy()  # (B, T, 3)
-                cls_logits = prediction['cls_logits'].cpu().numpy()
-                all_trajectories = prediction['all_trajectories'].cpu().numpy() # (B, K, T, 3)
+                poses = prediction['trajectory'].cpu().float().numpy()  # (B, T, 3)
+                cls_logits = prediction['cls_logits'].cpu().float().numpy()
+                all_trajectories = prediction['all_trajectories'].cpu().float().numpy() # (B, K, T, 3)
                 tokens = features['token']
 
                 if poses.shape[1] == 40:
